Fixed seat hold confirmation
I've moved the TicketSeatHold creation from the confirm view to the model
confirm method, to avoid confirm tickets and not reserve seats.

On confirm I'm creating a new TicketSeatHold of type 'R', I'm not removing
or changing the other existing TicketSeatHold of type 'H', 'P' or 'C',
because they're removed using the websocket server and when they're removed
the confirmed notification is sent to client, so the orange changes to red.